Destroy all handles on FdoD0ToD3
Since the StreamWorker holds a reference to the XENBUS_CONS interface,
xenbus will BUG_ON if the handles are not cleaned up before power down.
The service should close all handles in response to a
DBT_DEVICEQUERYREMOVE notification, but this may not be issued on
system power down.

Signed-off-by: Owen Smith <owen.smith@citrix.com>

Added missing UNREFERENCED_PARAMETER() to fix build failure.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>